# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clippzi API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Search/list users
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
  /users/{id}/followers:
    get:
      operationId: getUserFollowers
      tags: [users]
      summary: Get followers of a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}/following:
    get:
      operationId: getUserFollowing
      tags: [users]
      summary: Get users that a user follows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get user stats (followers, views, earnings)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
  /users/{id}/earnings:
    get:
      operationId: getUserEarnings
      tags: [users]
      summary: Get streamer earnings breakdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsSummary"
  /users/{id}/bank:
    get:
      operationId: getUserBankAccount
      tags: [users]
      summary: Get linked bank account info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
    post:
      operationId: linkBankAccount
      tags: [users]
      summary: Link a bank account for payouts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountInput"
      responses:
        "201":
          description: Linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"

  # ── POSTS ──
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        bio: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        bannerUrl: { type: ["string", "null"] }
        isVerified: { type: boolean }
        isLive: { type: boolean }
        followerCount: { type: integer }
        followingCount: { type: integer }
        postCount: { type: integer }
        totalViews: { type: integer }
        role: { type: string, enum: [user, streamer, admin] }
        createdAt: { type: string }
    UserInput:
      type: object
      required: [username, displayName, email]
      properties:
        username: { type: string }
        displayName: { type: string }
        email: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }
    UserUpdate:
      type: object
      properties:
        displayName: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }
        bannerUrl: { type: string }
        role: { type: string }
    UserStats:
      type: object
      required: [userId, followerCount, followingCount, postCount, totalViews, totalGiftsReceived, totalEarnings, totalGiftsSent]
      properties:
        userId: { type: integer }
        followerCount: { type: integer }
        followingCount: { type: integer }
        postCount: { type: integer }
        totalViews: { type: integer }
        totalGiftsReceived: { type: number }
        totalEarnings: { type: number }
        totalGiftsSent: { type: number }
    EarningsSummary:
      type: object
      required: [userId, totalGrossEarnings, streamerShare, platformShare, pendingPayout, paidOut, transactions]
      properties:
        userId: { type: integer }
        totalGrossEarnings: { type: number }
        streamerShare: { type: number }
        platformShare: { type: number }
        pendingPayout: { type: number }
        paidOut: { type: number }
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/GiftTransaction"
    BankAccount:
      type: object
      required: [id, userId, bankName, last4, status]
      properties:
        id: { type: integer }
        userId: { type: integer }
        bankName: { type: string }
        last4: { type: string }
        routingNumber: { type: ["string", "null"] }
        status: { type: string, enum: [pending, verified, failed] }
        createdAt: { type: string }
    BankAccountInput:
      type: object
      required: [bankName, accountNumber, routingNumber, accountHolderName]
      properties:
        bankName: { type: string }
        accountNumber: { type: string }
        routingNumber: { type: string }
        accountHolderName: { type: string }
    FollowInput:
      type: object
      required: [followerId, action]
      properties:
        followerId: { type: integer }
        action: { type: string, enum: [follow, unfollow] }
    FollowResult:
      type: object
      required: [following, followerCount]
      properties:
        following: { type: boolean }
        followerCount: { type: integer }

    # ── POST ──
...
    Gift:
      type: object
      required: [id, name, emoji, price, category, animationUrl]
      properties:
        id: { type: integer }
        name: { type: string }
        emoji: { type: string }
        price: { type: number }
        category: { type: string }
        animationUrl: { type: ["string", "null"] }
        iconUrl: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        rarity: { type: string, enum: [common, rare, epic, legendary] }
        isActive: { type: boolean }
    GiftTransaction:
      type: object
      required: [id, giftId, senderId, receiverId, amount, streamerShare, platformShare, createdAt]
      properties:
        id: { type: integer }
        giftId: { type: integer }
        gift: { $ref: "#/components/schemas/Gift" }
        senderId: { type: integer }
        sender: { $ref: "#/components/schemas/User" }
        receiverId: { type: integer }
        receiver: { $ref: "#/components/schemas/User" }
        streamId: { type: ["integer", "null"] }
        amount: { type: number }
        streamerShare: { type: number }
        platformShare: { type: number }
        quantity: { type: integer }
        message: { type: ["string", "null"] }
        createdAt: { type: string }
...