# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clippzi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  bannerUrl?: string;
  role?: string;
}